[android] adding dynamic filter menu action to SearchFragment toolbar
Here's a proof of concept attempt to re-draw the icon (Drawable) of the action.
For now it simply displays a monotonically increasing counter every time you touch it

I've made all the drawing code in a background thread, and once the drawable is ready the menuItem.setIcon() call is posted on the UI thread.

I tried using .recycle() on the Bitmap but this causes a crash saying that I can't draw on a recycled bitmap, so I guess that's unnecessary.
I suppose this Bitmap should only be recycled when the fragment is destroyed.